@lntricate1 2025-05-13
No worries. While working on this I realized that Base64 itself has not been updated in 5 years and I've been working on a PR for that. Hopefully when that's done, the changes I made here can be simplified and made more readable without losing performance.